+++ __tests__/useSwipeable.spec.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import { render, fireEvent, act } from "@testing-library/react";
+import { render, fireEvent, createEvent, act } from "@testing-library/react";
 import { useSwipeable } from "../src/index";
 import { LEFT, RIGHT, UP, DOWN, SwipeableProps } from "../src/types";
-import { expectSwipeFuncsDir } from "./helpers";
+import { expectSwipeFuncsDir, MockedSwipeFunctions } from "./helpers";
 
 const DIRECTIONS: [typeof LEFT, typeof RIGHT, typeof UP, typeof DOWN] = [
   LEFT,
@@ -11,14 +11,11 @@ const DIRECTIONS: [typeof LEFT, typeof RIGHT, typeof UP, typeof DOWN] = [
   DOWN,
 ];
 
-export type MockedSwipeFunctions = {
-  onSwiping: jest.Mock;
-  onSwiped: jest.Mock;
-  onSwipedLeft: jest.Mock;
-  onSwipedRight: jest.Mock;
-  onSwipedUp: jest.Mock;
-  onSwipedDown: jest.Mock;
-};
+const touchStart = "touchstart";
+const touchMove = "touchmove";
+const touchEnd = "touchend";
+const touchListeners = [touchStart, touchMove, touchEnd];
+
 function getMockedSwipeFunctions(): MockedSwipeFunctions {
   const onSwiped = "onSwiped";
   return DIRECTIONS.reduce(
@@ -36,12 +33,24 @@ const TESTING_TEXT = "touch here";
  */
 function SwipeableUsingHook({
   nodeName = "div",
+  onRefPassThrough,
   ...rest
-}: SwipeableProps & { nodeName?: string }) {
+}: SwipeableProps & {
+  nodeName?: string;
+  onRefPassThrough?(el: HTMLElement): void;
+}) {
   const eventHandlers = useSwipeable(rest);
   const Elem = nodeName as React.ElementType;
+
+  const refPassthrough = (el: HTMLElement) => {
+    onRefPassThrough?.(el);
+
+    // call useSwipeable ref prop with el
+    eventHandlers.ref(el);
+  };
+
   return (
-    <Elem {...eventHandlers}>
+    <Elem {...eventHandlers} ref={refPassthrough}>
       <span>{TESTING_TEXT}</span>
     </Elem>
   );
@@ -53,17 +62,36 @@ const MU = "mouseUp";
 const TS = "touchStart";
 const TM = "touchMove";
 const TE = "touchEnd";
+type touchTypes = typeof TS | typeof TM | typeof TE;
 
 const createClientXYObject = (x?: number, y?: number) => ({
   clientX: x,
   clientY: y,
 });
+type xyObj = { x?: number; y?: number };
 // Create touch event
-const cte = ({ x, y }: { x?: number; y?: number }) => ({
+const cte = ({ x, y }: xyObj) => ({
   touches: [createClientXYObject(x, y)],
 });
+// create touch event with timestamp
+const cteTs = ({
+  x,
+  y,
+  timeStamp,
+  type,
+  node,
+}: xyObj & { timeStamp?: number; type: touchTypes; node: HTMLElement }) => {
+  const e = createEvent[type](node, cte({ x, y }));
+  if (timeStamp) {
+    Object.defineProperty(e, "timeStamp", {
+      value: timeStamp,
+      writable: false,
+    });
+  }
+  return e;
+};
 // Create Mouse Event
-const cme = ({ x, y }: { x?: number; y?: number }) => ({
+const cme = ({ x, y }: xyObj) => ({
   ...createClientXYObject(x, y),
 });
 
@@ -151,14 +179,72 @@ describe("useSwipeable", () => {
     expect(onTap).not.toHaveBeenCalled();
   });
 
+  it("skips touch events with more than 1 touch", () => {
+    const swipeFuncs = getMockedSwipeFunctions();
+    const { getByText } = render(<SwipeableUsingHook {...swipeFuncs} />);
+
+    const touchArea = getByText(TESTING_TEXT);
+
+    const setupMultipleTouchEvent = (
+      touches: { x?: number; y?: number }[]
+    ) => ({
+      touches: touches.map((t) => createClientXYObject(t.x, t.y)),
+    });
+
+    fireEvent[TS](touchArea, cte({ x: 100, y: 10 }));
+    fireEvent[TM](
+      touchArea,
+      setupMultipleTouchEvent([
+        { x: 125, y: 0 },
+        { x: 130, y: 10 },
+      ])
+    );
+    fireEvent[TM](
+      touchArea,
+      setupMultipleTouchEvent([
+        { x: 130, y: 0 },
+        { x: 135, y: 10 },
+      ])
+    );
+    fireEvent[TE](touchArea, cte({}));
+
+    fireEvent[TS](
+      touchArea,
+      setupMultipleTouchEvent([
+        { x: 100, y: 0 },
+        { x: 110, y: 10 },
+      ])
+    );
+    fireEvent[TM](
+      touchArea,
+      setupMultipleTouchEvent([
+        { x: 125, y: 0 },
+        { x: 130, y: 10 },
+      ])
+    );
+    fireEvent[TM](
+      touchArea,
+      setupMultipleTouchEvent([
+        { x: 130, y: 0 },
+        { x: 135, y: 10 },
+      ])
+    );
+    fireEvent[TE](touchArea, cte({}));
+
+    expect(swipeFuncs.onSwiping).toHaveBeenCalledTimes(0);
+    expect(swipeFuncs.onSwiped).toHaveBeenCalledTimes(0);
+  });
+
   it("handles touch events and fires correct props with undefined values for config", () => {
     const swipeFuncs = getMockedSwipeFunctions();
     const undefinedConfigOptions: SwipeableProps = {
       delta: undefined,
-      preventDefaultTouchmoveEvent: undefined,
+      preventScrollOnSwipe: undefined,
       rotationAngle: undefined,
       trackMouse: undefined,
       trackTouch: undefined,
+      swipeDuration: undefined,
+      touchEventOptions: undefined,
     };
     const { getByText } = render(
       <SwipeableUsingHook {...swipeFuncs} {...undefinedConfigOptions} />
@@ -253,6 +339,54 @@ describe("useSwipeable", () => {
     );
   });
 
+  it("calls onTouchStartOrOnMouseDown and onTouchEndOrOnMouseUp for touch or mouse events", () => {
+    const onTouchStartOrOnMouseDownMock = jest.fn();
+    const onTouchEndOrOnMouseUpMock = jest.fn();
+    const { getByText, rerender } = render(
+      <SwipeableUsingHook
+        onTouchStartOrOnMouseDown={onTouchStartOrOnMouseDownMock}
+        onTouchEndOrOnMouseUp={onTouchEndOrOnMouseUpMock}
+        trackMouse={true}
+        trackTouch={false}
+      />
+    );
+
+    const touchArea = getByText(TESTING_TEXT);
+
+    fireEvent[MD](touchArea, cme({ x: 100, y: 100 }));
+    fireEvent[MM](touchArea, cme({ x: 125, y: 100 }));
+    fireEvent[MU](document, cme({}));
+
+    expect(onTouchStartOrOnMouseDownMock).toHaveBeenCalled();
+    expect(onTouchStartOrOnMouseDownMock.mock.calls[0][0].event.type).toEqual(
+      "mousedown"
+    );
+    expect(onTouchEndOrOnMouseUpMock).toHaveBeenCalled();
+    expect(onTouchEndOrOnMouseUpMock.mock.calls[0][0].event.type).toEqual(
+      "mouseup"
+    );
+
+    rerender(
+      <SwipeableUsingHook
+        onTouchStartOrOnMouseDown={onTouchStartOrOnMouseDownMock}
+        onTouchEndOrOnMouseUp={onTouchEndOrOnMouseUpMock}
+      />
+    );
+
+    fireEvent[TS](touchArea, cte({ x: 100, y: 100 }));
+    fireEvent[TM](touchArea, cte({ x: 100, y: 125 }));
+    fireEvent[TE](touchArea, cte({}));
+
+    expect(onTouchStartOrOnMouseDownMock).toHaveBeenCalledTimes(2);
+    expect(onTouchStartOrOnMouseDownMock.mock.calls[1][0].event.type).toEqual(
+      "touchstart"
+    );
+    expect(onTouchEndOrOnMouseUpMock).toHaveBeenCalledTimes(2);
+    expect(onTouchEndOrOnMouseUpMock.mock.calls[1][0].event.type).toEqual(
+      "touchend"
+    );
+  });
+
   it("correctly calls onSwipeStart for first swipe event", () => {
     const onSwipeStart = jest.fn();
     const { getByText } = render(
@@ -280,17 +414,70 @@ describe("useSwipeable", () => {
     expect(onSwipeStart).toHaveBeenCalledTimes(2);
   });
 
-  it("calls preventDefault when swiping in direction that has a callback", () => {
-    const onSwipedDown = jest.fn();
-
-    const { getByText } = render(
+  it("calls callbacks appropriately for swipeDuration", () => {
+    const onSwiped = jest.fn();
+    const onSwiping = jest.fn();
+    const { getByText, rerender } = render(
       <SwipeableUsingHook
-        onSwipedDown={onSwipedDown}
-        preventDefaultTouchmoveEvent
+        onSwiped={onSwiped}
+        onSwiping={onSwiping}
+        swipeDuration={15}
       />
     );
 
+    const el = getByText(TESTING_TEXT);
+    const fE = fireEvent;
+
+    fE(el, cteTs({ x: 100, y: 100, node: el, type: TS, timeStamp: 1000 }));
+    fE(el, cteTs({ x: 100, y: 125, node: el, type: TM, timeStamp: 1010 }));
+    fE(el, cteTs({ x: 100, y: 150, node: el, type: TM, timeStamp: 1020 }));
+    fE(el, cteTs({ x: 100, y: 175, node: el, type: TM, timeStamp: 1030 }));
+    fE(el, cteTs({ node: el, type: TE, timeStamp: 1040 }));
+
+    expect(onSwiping).toHaveBeenCalledTimes(1);
+    expect(onSwiped).not.toHaveBeenCalled();
+
+    onSwiping.mockClear();
+    onSwiped.mockClear();
+    rerender(
+      <SwipeableUsingHook
+        onSwiped={onSwiped}
+        onSwiping={onSwiping}
+        swipeDuration={50}
+      />
+    );
+
+    fE(el, cteTs({ x: 100, y: 100, node: el, type: TS, timeStamp: 1000 }));
+    fE(el, cteTs({ x: 100, y: 125, node: el, type: TM, timeStamp: 1010 }));
+    fE(el, cteTs({ x: 100, y: 150, node: el, type: TM, timeStamp: 1020 }));
+    fE(el, cteTs({ x: 100, y: 175, node: el, type: TM, timeStamp: 1030 }));
+    fE(el, cteTs({ node: el, type: TE, timeStamp: 1040 }));
+
+    expect(onSwiping).toHaveBeenCalledTimes(3);
+    expect(onSwiped).toHaveBeenCalled();
+  });
+
+  it("calls preventDefault when swiping in direction with callback defined", () => {
+    const onSwipedDown = jest.fn();
+
+    const { getByText, rerender } = render(
+      <SwipeableUsingHook onSwipedDown={undefined} preventScrollOnSwipe />
+    );
+
     const touchArea = getByText(TESTING_TEXT);
+
+    fireEvent[TS](touchArea, cte({ x: 100, y: 100 }));
+    fireEvent[TM](touchArea, cte({ x: 100, y: 150 }));
+    fireEvent[TM](touchArea, cte({ x: 100, y: 200 }));
+    fireEvent[TE](touchArea, cte({}));
+
+    // Validate `undefined` does not trigger defaultPrevented
+    expect(onSwipedDown).not.toHaveBeenCalled();
+    expect(defaultPrevented).toBe(0);
+
+    rerender(
+      <SwipeableUsingHook onSwipedDown={onSwipedDown} preventScrollOnSwipe />
+    );
 
     fireEvent[TS](touchArea, cte({ x: 100, y: 100 }));
     fireEvent[TM](touchArea, cte({ x: 100, y: 125 }));
@@ -328,7 +515,7 @@ describe("useSwipeable", () => {
     const onSwiped = jest.fn();
 
     const { getByText, rerender } = render(
-      <SwipeableUsingHook onSwiping={onSwiping} preventDefaultTouchmoveEvent />
+      <SwipeableUsingHook onSwiping={onSwiping} preventScrollOnSwipe />
     );
 
     const touchArea = getByText(TESTING_TEXT);
@@ -340,9 +527,7 @@ describe("useSwipeable", () => {
     expect(onSwiping).toHaveBeenCalled();
     expect(defaultPrevented).toBe(1);
 
-    rerender(
-      <SwipeableUsingHook onSwiped={onSwiped} preventDefaultTouchmoveEvent />
-    );
+    rerender(<SwipeableUsingHook onSwiped={onSwiped} preventScrollOnSwipe />);
 
     fireEvent[TS](touchArea, cte({ x: 100, y: 100 }));
     fireEvent[TM](touchArea, cte({ x: 100, y: 50 }));
@@ -350,6 +535,174 @@ describe("useSwipeable", () => {
 
     expect(onSwiped).toHaveBeenCalled();
     expect(defaultPrevented).toBe(2);
+  });
+
+  it("calls preventDefault appropriately when preventScrollOnSwipe value changes", () => {
+    const onSwipedDown = jest.fn();
+
+    const { getByText, rerender } = render(
+      <SwipeableUsingHook onSwipedDown={onSwipedDown} preventScrollOnSwipe />
+    );
+
+    const touchArea = getByText(TESTING_TEXT);
+
+    fireEvent[TS](touchArea, cte({ x: 100, y: 100 }));
+    fireEvent[TM](touchArea, cte({ x: 100, y: 125 }));
+    fireEvent[TM](touchArea, cte({ x: 100, y: 150 }));
+
+    // change preventScrollOnSwipe in middle of swipe
+    rerender(
+      <SwipeableUsingHook
+        onSwipedDown={onSwipedDown}
+        preventScrollOnSwipe={false}
+      />
+    );
+
+    fireEvent[TM](touchArea, cte({ x: 100, y: 175 }));
+    fireEvent[TM](touchArea, cte({ x: 100, y: 200 }));
+    fireEvent[TE](touchArea, cte({}));
+
+    expect(onSwipedDown).toHaveBeenCalled();
+    expect(defaultPrevented).toBe(2);
+  });
+
+  it("defaults touchEventOptions passive", () => {
+    const onSwiping = jest.fn();
+
+    let addEventListenerSpy: jest.SpyInstance | undefined;
+    const onRefPassThrough = (el: HTMLElement) => {
+      if (el === null) return;
+      // re-assign to element and spy
+      addEventListenerSpy = jest.spyOn(el, "addEventListener");
+    };
+
+    render(
+      <SwipeableUsingHook
+        onSwiping={onSwiping}
+        onRefPassThrough={onRefPassThrough}
+      />
+    );
+
+    // NOTE: this is what addEventListenerSpy.mock.calls looks like:
+    // [
+    //   [ 'touchstart', [Function: onStart], { passive: true } ],
+    //   [ 'touchmove', [Function: onMove], { passive: true } ],
+    //   [ 'touchend', [Function: onEnd], { passive: true } ]
+    // ]
+    expect(addEventListenerSpy?.mock.calls.length).toBe(3);
+    const calls = addEventListenerSpy?.mock.calls || [];
+    touchListeners.forEach((l, idx) => {
+      const [arg1, , arg3] = calls[idx] as any[];
+      expect(arg1).toBe(l);
+      expect(arg3).toEqual({ passive: true });
+    });
+
+    expect(onSwiping).not.toHaveBeenCalled();
+    expect(defaultPrevented).toBe(0);
+  });
+
+  it("preventScrollOnSwipe overwrites touchEventOptions passive", () => {
+    const onSwiping = jest.fn();
+
+    let addEventListenerSpy: jest.SpyInstance | undefined;
+    const onRefPassThrough = (el: HTMLElement) => {
+      if (el === null) return;
+      // re-assign to element and spy
+      addEventListenerSpy = jest.spyOn(el, "addEventListener");
+    };
+
+    render(
+      <SwipeableUsingHook
+        onSwiping={onSwiping}
+        onRefPassThrough={onRefPassThrough}
+        preventScrollOnSwipe
+      />
+    );
+
+    expect(addEventListenerSpy?.mock.calls.length).toBe(3);
+    const calls = addEventListenerSpy?.mock.calls || [];
+    touchListeners.forEach((l, idx) => {
+      const [arg1, , arg3] = calls[idx] as any[];
+      expect(arg1).toBe(l);
+      if (l === touchMove) {
+        // preventScrollOnSwipe overrides passive for touchmove
+        expect(arg3).toEqual({ passive: false });
+      } else {
+        expect(arg3).toEqual({ passive: true });
+      }
+    });
+
+    expect(onSwiping).not.toHaveBeenCalled();
+    expect(defaultPrevented).toBe(0);
+  });
+
+  it("changing touchEventOptions.passive re-attaches event listeners", () => {
+    const onSwiping = jest.fn();
+
+    // set to document to start to avoid TS issues
+    let addEventListenerSpy = jest.spyOn(document, "addEventListener");
+    const onRefPassThrough = (el: HTMLElement) => {
+      if (el === null) return;
+      // re-assign to element and spy
+      addEventListenerSpy = jest.spyOn(el, "addEventListener");
+    };
+
+    const { rerender } = render(
+      <SwipeableUsingHook
+        onSwiping={onSwiping}
+        onRefPassThrough={onRefPassThrough}
+      />
+    );
+
+    expect(addEventListenerSpy.mock.calls.length).toBe(3);
+    let calls = addEventListenerSpy.mock.calls;
+    touchListeners.forEach((l, idx) => {
+      const call: any = calls[idx];
+      expect(call[0]).toBe(l);
+      expect(call[2]).toEqual({ passive: true });
+    });
+
+    expect(onSwiping).not.toHaveBeenCalled();
+    expect(defaultPrevented).toBe(0);
+
+    // reset spy before re-render
+    addEventListenerSpy.mockClear();
+
+    // change touchEventOptions.passive from default
+    rerender(
+      <SwipeableUsingHook
+        onSwiping={onSwiping}
+        onRefPassThrough={onRefPassThrough}
+        touchEventOptions={{ passive: false }}
+      />
+    );
+
+    expect(addEventListenerSpy.mock.calls.length).toBe(3);
+    calls = addEventListenerSpy.mock.calls;
+    touchListeners.forEach((l, idx) => {
+      const call: any = calls[idx];
+      expect(call[0]).toBe(l);
+      expect(call[2]).toEqual({ passive: false });
+    });
+
+    expect(onSwiping).not.toHaveBeenCalled();
+    expect(defaultPrevented).toBe(0);
+  });
+
+  it("does not fire onSwiped when under delta", () => {
+    const onSwiped = jest.fn();
+    const { getByText } = render(
+      <SwipeableUsingHook onSwiped={onSwiped} delta={40} />
+    );
+
+    const touchArea = getByText(TESTING_TEXT);
+
+    fireEvent[TS](touchArea, cte({ x: 100, y: 100 }));
+    fireEvent[TM](touchArea, cte({ x: 120, y: 100 }));
+    fireEvent[TM](touchArea, cte({ x: 130, y: 100 }));
+    fireEvent[TE](touchArea, cte({}));
+
+    expect(onSwiped).not.toHaveBeenCalled();
   });
 
   it("does not re-check delta when swiping already in progress", () => {
@@ -715,5 +1068,29 @@ describe("useSwipeable", () => {
     pages.forEach((page, idx) => {
       expect(onSwipedSpy.mock.calls[idx][0]).toBe(page);
     });
+  });
+
+  it("Does not trigger onSwiped again clicking document after a previous touch swipe when trackTouch and trackMouse are present", () => {
+    // See issue #304
+    // - https://github.com/FormidableLabs/react-swipeable/issues/304
+    const onSwiped = jest.fn();
+
+    const { getByText } = render(
+      <SwipeableUsingHook onSwiped={onSwiped} trackMouse trackTouch />
+    );
+
+    const touchArea = getByText(TESTING_TEXT);
+
+    fireEvent[TS](touchArea, cte({ x: 100, y: 100 }));
+    fireEvent[TM](touchArea, cte({ x: 100, y: 200 }));
+    fireEvent[TE](touchArea, cte({}));
+
+    expect(onSwiped).toHaveBeenCalledTimes(1);
+
+    fireEvent[MD](document, cme({ x: 100, y: 100 }));
+    fireEvent[MM](document, cme({ x: 100, y: 100 }));
+    fireEvent[MU](document, cme({}));
+    // verify we did NOT trigger another swipe
+    expect(onSwiped).toHaveBeenCalledTimes(1);
   });
 });